Lab - 3
Review any feedback from lab week 2 and adjust your code.  This includes useful validation of all arguments used in the constructor and set methods
add a private instance variable to store a list of interests.  This could be an ArrayList of type String
add a method to add interests for a Student.  The method should validate that the interest comes from a predefined list.  For example, in the Card class, we ensured that a Card object would have a suit that is either "hearts","diamonds","clubs", or "spades".  You should have a list of things like "hiking","skiing","reading","coding","gaming", etc...